+++ src/ATAP.Utilities.Security.Secrets.PowerShell/ATAP.Utilities.Security.Secrets.PowerShell.psm1
@@ -15,6 +15,7 @@ Set-Alias -Name Sync-DedicatedSecrets -Value Sync-BitWardenDedicatedSecrets
 
 Export-ModuleMember -Function @(
   'Get-BitWardenCredential',
+  'Invoke-RotateSecretsATAP',
   'List-BitwardenSecrets',
   'Load-BitwardenBackup',
   'New-BitwardenBackup',