+++ src/ATAP.Utilities.Security.Secrets.PowerShell/public/List-BitwardenSecrets.ps1
@@ -69,7 +69,7 @@ function List-BitwardenSecrets {
 
   begin {
     $fn = 'List-BitwardenSecrets'
-    $mn = 'ATAP.Utilities.Security.Powershell'
+    $mn = 'ATAP.Utilities.Security.Secrets.PowerShell'
 
     Write-PSFMessage -FunctionName $fn -ModuleName $mn -Level Debug -Message "Entering function $fn in module $mn"