+++ src/ATAP.Utilities.Security.Secrets.PowerShell/public/New-BitwardenBackup.ps1
@@ -87,7 +87,7 @@ function New-BitwardenBackup {
 
   begin {
     $fn = 'New-BitwardenBackup'
-    $mn = 'ATAP.Utilities.Security.Powershell'
+    $mn = 'ATAP.Utilities.Security.Secrets.PowerShell'
 
     Write-PSFMessage -FunctionName $fn -ModuleName $mn -Level Debug -Message "Entering function $fn in module $mn"